Modularizes admin export to simplify future maintenance
