+++ samples/packages/admin-export/src/main/kotlin/PackageConfig.kt
@@ -45,7 +45,7 @@ object PackageConfig : CustomPackage(
     ),
     containerImage = "ghcr.io/atlanhq/csa-admin-export:${Atlan.VERSION}",
     containerImagePullPolicy = "Always",
-    containerCommand = listOf("/dumb-init", "--", "java", "AdminExporter"),
+    containerCommand = listOf("/dumb-init", "--", "java", "com.atlan.pkg.ae.AdminExporter"),
     outputs = WorkflowOutputs(
         mapOf(
             "debug-logs" to "/tmp/debug.log",
