+++ samples/packages/admin-export/src/test/kotlin/ExportAllAdminInfoTest.kt
@@ -1,6 +1,7 @@
 /* SPDX-License-Identifier: Apache-2.0
    Copyright 2023 Atlan Pte. Ltd. */
 import com.atlan.pkg.PackageTest
+import com.atlan.pkg.ae.AdminExporter
 import com.atlan.pkg.serde.xls.ExcelReader
 import org.testng.Assert.assertFalse
 import org.testng.Assert.assertTrue
@@ -31,7 +32,7 @@ class ExportAllAdminInfoTest : PackageTest() {
                     "purposes",
                     "policies",
                 ),
-                includeNativePolicies = true,
+                includeNativePolicies = false,
             ),
         )
         AdminExporter.main(arrayOf(testDirectory))
