+++ samples/packages/admin-export/src/test/kotlin/ExportUsersTest.kt
@@ -1,6 +1,7 @@
 /* SPDX-License-Identifier: Apache-2.0
    Copyright 2023 Atlan Pte. Ltd. */
 import com.atlan.pkg.PackageTest
+import com.atlan.pkg.ae.AdminExporter
 import com.atlan.pkg.serde.xls.ExcelReader
 import org.testng.Assert.assertFalse
 import org.testng.Assert.assertTrue
