+++ settings.gradle.kts
@@ -8,6 +8,7 @@ include("package-toolkit:config")
 include("package-toolkit:events")
 include("package-toolkit:runtime")
 include("package-toolkit:testing")
+include("samples:packages:admin-export")
 include("samples:packages:api-token-connection-admin")
 include("samples:packages:asset-export-basic")
 include("samples:packages:asset-import")
